+++ R/type-rcrd.R
@@ -89,11 +89,11 @@ vec_cast.vctrs_rcrd <- function(x, to, ...) UseMethod("vec_cast.vctrs_rcrd")
 
 #' @method vec_cast.vctrs_rcrd vctrs_rcrd
 #' @export
-vec_cast.vctrs_rcrd.vctrs_rcrd <- function(x, to, ...) {
+vec_cast.vctrs_rcrd.vctrs_rcrd <- function(x, to, ..., x_arg = x_arg, to_arg = to_arg) {
   # This assumes that we don't have duplicate field names,
   # which is verified even in the constructor.
   if (!setequal(fields(x), fields(to))) {
-    stop_incompatible_cast(x, to)
+    stop_incompatible_cast(x, to, x_arg = x_arg, to_arg = to_arg)
   }
 
   new_data <- map2(